+++ src/MessSharp/Model/ISourceFileParser.cs
@@ -1,0 +1,6 @@
+namespace MessSharp.Model;
+
+public interface ISourceFileParser
+{
+    SourceFile ParseFile(string path);
+}