+++ src/MessSharp/Runner/IRunner.cs
@@ -1,0 +1,8 @@
+using RuleSetType = MessSharp.Rule.RuleSet;
+
+namespace MessSharp.Runner;
+
+public interface IRunner
+{
+    MessSharp.Report.Report Run(RunOptions opts);
+}